requirements.txt
requests~=2.31.0
mysql-connector-python~=8.0.31
envparse~=0.2.0
python-dotenv~=1.0.0
redis~=4.5.5
